Move utils to commons folder
