+++ include/seahowl/aero/tower_aero.h
@@ -4,7 +4,7 @@
 
 #include <seahowl/aero/reference_point_aero.h>
 #include <seahowl/aero/wind_models.h>
-#include <seahowl/core/utils.h>
+#include <seahowl/commons/utils.h>
 #include <seahowl/commons/numerics.h>
 
 namespace seahowl {
