+++ src/core/blade.cpp
@@ -3,7 +3,7 @@
 #include <seahowl/elasto/reference_point_elasto.h>
 #include <seahowl/elasto/blade_elasto.h>
 #include <seahowl/aero/blade_aero.h>
-#include <seahowl/core/utils.h>
+#include <seahowl/commons/utils.h>
 
 #include <memory>
